Move tests from `test_datatypes` into `test_datamodel`
There is no need to have these two separate test modules. Tests from
`test_datatypes` were moved into `test_datamodel` with the ordering
of test cases rearranged.

The old test module `test_datatypes` was deleted.
