+++ tests/test_datamodel.py
@@ -42,18 +42,6 @@ TESTDIR = os.path.abspath(os.path.dirname(__file__))
 TEST_YAML = os.path.join(TESTDIR, "data", "test.yml")
 
 
-class DataModelTestCase(unittest.TestCase):
-    """Test validation of a test schema from a YAML file."""
-
-    schema_obj: Schema
-
-    def test_validation(self) -> None:
-        """Load test file and validate it using the data model."""
-        with open(TEST_YAML) as test_yaml:
-            data = yaml.safe_load(test_yaml)
-            self.schema_obj = Schema.model_validate(data)
-
-
 class ColumnTestCase(unittest.TestCase):
     """Test the `Column` class."""
 
@@ -235,6 +223,37 @@ class ColumnTestCase(unittest.TestCase):
         Column(**bool_coldata)
 
 
+class TableTestCase(unittest.TestCase):
+    """Test the `Table` class."""
+
+    def test_validation(self) -> None:
+        # Default initialization should throw an exception.
+        with self.assertRaises(ValidationError):
+            Table()
+
+        # Setting only name should throw an exception.
+        with self.assertRaises(ValidationError):
+            Table(name="testTable")
+
+        # Setting name and id should throw an exception from missing columns.
+        with self.assertRaises(ValidationError):
+            Index(name="testTable", id="#test_id")
+
+        testCol = Column(name="testColumn", id="#test_id", datatype="string", length=256)
+
+        # Setting name, id, and columns should not throw an exception and
+        # should load data correctly.
+        tbl = Table(name="testTable", id="#test_id", columns=[testCol])
+        self.assertEqual(tbl.name, "testTable", "name should be 'testTable'")
+        self.assertEqual(tbl.id, "#test_id", "id should be '#test_id'")
+        self.assertEqual(tbl.columns, [testCol], "columns should be ['testColumn']")
+
+        # Creating a table with duplicate column names should raise an
+        # exception.
+        with self.assertRaises(ValidationError):
+            Table(name="testTable", id="#test_id", columns=[testCol, testCol])
+
+
 class ConstraintTestCase(unittest.TestCase):
     """Test the `UniqueConstraint`, `Index`, `CheckConstraint`, and
     `ForeignKeyConstraint` classes.
@@ -376,37 +395,6 @@ class ConstraintTestCase(unittest.TestCase):
         self.assertEqual(col.expression, "1+2", "expression should be '1+2'")
 
 
-class TableTestCase(unittest.TestCase):
-    """Test the `Table` class."""
-
-    def test_validation(self) -> None:
-        # Default initialization should throw an exception.
-        with self.assertRaises(ValidationError):
-            Table()
-
-        # Setting only name should throw an exception.
-        with self.assertRaises(ValidationError):
-            Table(name="testTable")
-
-        # Setting name and id should throw an exception from missing columns.
-        with self.assertRaises(ValidationError):
-            Index(name="testTable", id="#test_id")
-
-        testCol = Column(name="testColumn", id="#test_id", datatype="string", length=256)
-
-        # Setting name, id, and columns should not throw an exception and
-        # should load data correctly.
-        tbl = Table(name="testTable", id="#test_id", columns=[testCol])
-        self.assertEqual(tbl.name, "testTable", "name should be 'testTable'")
-        self.assertEqual(tbl.id, "#test_id", "id should be '#test_id'")
-        self.assertEqual(tbl.columns, [testCol], "columns should be ['testColumn']")
-
-        # Creating a table with duplicate column names should raise an
-        # exception.
-        with self.assertRaises(ValidationError):
-            Table(name="testTable", id="#test_id", columns=[testCol, testCol])
-
-
 class SchemaTestCase(unittest.TestCase):
     """Test the `Schema` class."""
 
@@ -477,6 +465,12 @@ class SchemaTestCase(unittest.TestCase):
         with self.assertRaises(KeyError):
             # Test that an invalid id raises an exception.
             sch["#bad_id"]
+
+    def test_model_validate(self) -> None:
+        """Load test file and validate it using the data model."""
+        with open(TEST_YAML) as test_yaml:
+            data = yaml.safe_load(test_yaml)
+            Schema.model_validate(data)
 
 
 class SchemaVersionTest(unittest.TestCase):
@@ -608,6 +602,91 @@ class ValidationFlagsTest(unittest.TestCase):
         schema_dict["tables"][0]["description"] = "Test table"
         schema_dict["tables"][0]["columns"][0]["description"] = "Test column"
         Schema.model_validate(schema_dict, context=cxt)
+
+
+class RedundantDatatypesTest(unittest.TestCase):
+    """Test validation of redundant datatype definitions."""
+
+    def test_mysql_datatypes(self) -> None:
+        class ColumnGenerator:
+            """Generate column data for redundant datatype testing."""
+
+            def __init__(self, name, id, db_name):
+                self.name = name
+                self.id = id
+                self.db_name = db_name
+                self.context = {"check_redundant_datatypes": True}
+
+            def col(self, datatype: str, db_datatype: str, length=None):
+                return Column.model_validate(
+                    {
+                        "name": self.name,
+                        "@id": self.id,
+                        "datatype": datatype,
+                        f"{self.db_name}:datatype": db_datatype,
+                        "length": length,
+                    },
+                    context=self.context,
+                )
+
+        """Test that redundant datatype definitions raise an error."""
+        coldata = ColumnGenerator("test_col", "#test_col_id", "mysql")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("double", "DOUBLE")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("int", "INTEGER")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("float", "FLOAT")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("char", "CHAR", length=8)
+
+        with self.assertRaises(ValidationError):
+            coldata.col("string", "VARCHAR", length=32)
+
+        with self.assertRaises(ValidationError):
+            coldata.col("byte", "TINYINT")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("short", "SMALLINT")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("long", "BIGINT")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("boolean", "BOOLEAN")
+
+        with self.assertRaises(ValidationError):
+            coldata.col("unicode", "NVARCHAR", length=32)
+
+        with self.assertRaises(ValidationError):
+            coldata.col("timestamp", "TIMESTAMP")
+
+        # DM-42257: Felis does not handle unbounded text types properly.
+        # coldata.col("text", "TEXT", length=32)
+
+        with self.assertRaises(ValidationError):
+            coldata.col("binary", "LONGBLOB", length=1024)
+
+        with self.assertRaises(ValidationError):
+            # Same type and length
+            coldata.col("string", "VARCHAR(128)", length=128)
+
+        # Check the old type mapping for MySQL, which is now okay
+        coldata.col("boolean", "BIT(1)")
+
+        # Different types, which is okay
+        coldata.col("double", "FLOAT")
+
+        # Same base type with different lengths, which is okay
+        coldata.col("string", "VARCHAR(128)", length=32)
+
+        # Different string types, which is okay
+        coldata.col("string", "CHAR", length=32)
+        coldata.col("unicode", "CHAR", length=32)
 
 
 if __name__ == "__main__":
